streaming

+++ script/deadlock_performance.sh
@@ -1,17 +1,19 @@
 #! /bin/bash
 
+DEBUG=false
+
 ################ System Variables ################ 
 LOG_DIR=../data/deadlock_Performance_pipeline_
-TEST_FILE=../microbench/instantSubmission/pipeline.py
+TEST_FILE=../microbench/instantSubmission/shuffle.py
 OBJECT_STORE_SIZE=4000000000
 OBJECT_SIZE=100000000
 
 ################ Test Techniques ################ 
-Production_RAY=true
+Production_RAY=false
 DFS=false
 CONSTANT_WAIT=false
 DEADLOCK_ONE=false
-DEADLOCK_TWO=false
+DEADLOCK_TWO=true
 
 function Test()
 {
@@ -19,17 +21,29 @@ function Test()
 	DEADLOCK1=$2
 	DEADLOCK2=$3
 	RESULT_FILE=$LOG_DIR$4.csv
-	test -f "$RESULT_FILE" && rm $RESULT_FILE
-	echo "std,var,working_set,object_store_size,object_size,time" >>$RESULT_FILE
-	for w in {1,2,4,8}
-	do
-		let a=2000
-		echo $4 -w $w wait time $a
-		#RAY_BACKEND_LOG_LEVEL=debug \
-		RAY_object_spilling_threshold=1.0 RAY_spill_wait_time=$a RAY_enable_BlockTasksSpill=$BLOCKSPILL  RAY_enable_Deadlock1=$DEADLOCK1 RAY_enable_Deadlock2=$DEADLOCK2 \
-		python $TEST_FILE -w $w -r $RESULT_FILE -o $OBJECT_STORE_SIZE -os $OBJECT_SIZE -t 10
-	done
-	rm -rf /tmp/ray/*
+	if $DEBUG;
+	then
+		#for w in {1,2,4,8}
+		#do
+		w=8
+			let a=20000
+			echo $4 -w $w wait time $a
+			RAY_BACKEND_LOG_LEVEL=debug \
+			RAY_object_spilling_threshold=1.0 RAY_spill_wait_time=$a RAY_enable_BlockTasksSpill=$BLOCKSPILL  RAY_enable_Deadlock1=$DEADLOCK1 RAY_enable_Deadlock2=$DEADLOCK2 \
+			python $TEST_FILE -w $w -o $OBJECT_STORE_SIZE -os $OBJECT_SIZE -t 1
+		#done
+	else
+		test -f "$RESULT_FILE" && rm $RESULT_FILE
+		echo "std,var,working_set,object_store_size,object_size,time" >>$RESULT_FILE
+		for w in {1,2,4,8}
+		do
+			let a=20000
+			echo $4 -w $w wait time $a
+			RAY_object_spilling_threshold=1.0 RAY_spill_wait_time=$a RAY_enable_BlockTasksSpill=$BLOCKSPILL  RAY_enable_Deadlock1=$DEADLOCK1 RAY_enable_Deadlock2=$DEADLOCK2 \
+			python $TEST_FILE -w $w -r $RESULT_FILE -o $OBJECT_STORE_SIZE -os $OBJECT_SIZE -t 5
+		done
+		rm -rf /tmp/ray/*
+	fi
 }
 
 if $Production_RAY;
